toastの処理を修正
